remove code from get_tasks
