+++ mteb/abstasks/abstask.py
@@ -20,6 +20,7 @@ from tqdm.auto import tqdm
 
 from mteb._hf_integration.eval_model import HFEvalMeta, HFEvalTaskConfig
 from mteb._hf_integration.hf_hub_utils import _get_file_on_hub
+from mteb._log_once import LogOnce
 from mteb._set_seed import _set_seed
 from mteb.languages import LanguageScripts
 from mteb.models import (
@@ -30,7 +31,7 @@ from mteb.models import (
 from mteb.timing import TimingStack
 
 if TYPE_CHECKING:
-    from collections.abc import Mapping
+    from collections.abc import Iterator, Mapping
 
     from typing_extensions import Self
 
@@ -43,6 +44,7 @@ if TYPE_CHECKING:
 
 
 logger = logging.getLogger(__name__)
+log_once = LogOnce(logger)
 
 
 def _multilabel_subsampling(
@@ -807,3 +809,53 @@ class AbsTask(ABC):  # noqa: PLR0904
     def superseded_by(self) -> str | None:
         """If the dataset is superseded by another dataset, return the name of the new dataset."""
         return self.metadata.superseded_by
+
+
+def get_abstask_prompt(task_name: str) -> str:
+    """Get the prompt defined by the abstask class (`AbsTask.abstask_prompt`) of a task.
+
+    The task is resolved from the imported task classes instead of the task registry, such that it
+    doesn't only work for tasks within mteb, but also for tasks outside of it, e.g. mock tasks or
+    user defined tasks.
+
+    Args:
+        task_name: The name of the task, e.g. "SciFact".
+
+    Returns:
+        The prompt defined by the abstask class of the task, or an empty string if its abstask class
+            doesn't define one.
+    """
+    if task_name not in _task_name_to_prompt:
+        prompt = ""
+        for task_cls in _iter_task_subclasses():
+            # `abstask_prompt` is defined on the abstask class, e.g. `AbsTaskRetrieval`, and is only
+            # annotated on `AbsTask`, so abstasks such as `AbsTaskAggregate` don't define one
+            if task_cls.metadata.name == task_name:
+                prompt = getattr(task_cls, "abstask_prompt", "")
+                if prompt:
+                    break
+        _task_name_to_prompt[task_name] = prompt
+
+    prompt = _task_name_to_prompt[task_name]
+    if not prompt:
+        log_once.warning(
+            f"No prompt found on the abstask class of task '{task_name}'. Using an empty prompt."
+        )
+    return prompt
+
+
+def _iter_task_subclasses() -> Iterator[type[AbsTask]]:
+    """Iterate over the imported subclasses of `AbsTask` that define a metadata."""
+    stack: list[Any] = [AbsTask]
+    seen: set[Any] = set()
+    while stack:
+        for subclass in stack.pop().__subclasses__():
+            if subclass in seen:
+                continue
+            seen.add(subclass)
+            stack.append(subclass)
+            if getattr(subclass, "metadata", None) is not None:
+                yield subclass
+
+
+_task_name_to_prompt: dict[str, str] = {}
